beta

+++ bestPayApp/models.py
@@ -50,7 +50,7 @@ class Payment(models.Model):
         super().save(*args, **kwargs)
 
     def amount_value(self):
-        return int(self.amount) * 100
+        return float(self.amount) * 100
 
     def verify_payment(self):
         paystack = Paystack()
